pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "disco_rl_project"
version = "0.1.0"
description = "Meta-training and evaluation using Disco RL"
requires-python = ">=3.9"
dependencies = [
    "jax",
    "jaxlib",
    "dm-haiku",
    "optax",
    "rlax",
    "distrax",
    "chex",
    "ml_collections",
    "pandas",
    "seaborn",
    "tqdm",
    "requests",
    "matplotlib",
    "numpy"
]

[tool.setuptools.packages.find]
where = ["."]